Style setup, demo component.
Setup sass files to use styles across project. Site-wide styling is
contained in the global.scss file, and specific styling per component is
stored in the same folder as the component, as the same name, but with a
.sass extension. Variables are stored in the _variables file. The Demo
component performs initial query on load (useEffect) and passes state
down to the other components (eventually ChartJS) and currently
React-JSON, which is another library that lets us open up our JSON
response data and styles it.

The NavBar is routing the user around teh site, but has not yet been
styled. Rendered at every page. React router is selectively rendering
components depending on where the user navigates in our application.